removing all streaming code
