+++ colossus/src/main/scala/colossus/controller/OutputController.scala
@@ -9,42 +9,6 @@ import scala.util.{Success, Failure}
 
 import service.{NotConnectedException, RequestTimeoutException}
 
-/**
- * The DataQueue is used only when processing a stream.  It is used to buffer
- * data in between when the stream produces it and when it is ready to write to
- * the output buffer.  It also ensures that we don't pull too much data from a
- * stream, for example if a stream is infinite
- */
-class DataQueue(maxBytes: Long) {
-
-  private var total = 0L
-  private val queue = new LinkedList[(DataBuffer, Long)]
-
-  def dataSize = total
-  def itemSize = queue.size
-
-  def isFull = total >= maxBytes
-  def isEmpty = itemSize == 0
-
-  def head = queue.peek._1
-
-  /**
-   * returns true if the queue can accept more data
-   */
-  def enqueue(data: DataBuffer): Boolean = {
-    val size = data.remaining
-    total += size
-    queue.add(data -> size)
-    isFull
-  }
-
-  def dequeue: DataBuffer = {
-    val (data, size) = queue.remove
-    total -= size
-    data
-  }
-
-}
 
 case class QueuedItem[T](item: T, postWrite: QueuedItem.PostWrite, creationTimeMillis: Long) {
   def isTimedOut(now: Long, timeout: Duration) = timeout.isFinite && now > (creationTimeMillis + timeout.toMillis)
@@ -72,34 +36,6 @@ class MessageQueue[T](maxSize: Int) {
 
 }
 
-
-sealed trait OutputState {
-  def canPush : Boolean
-}
-sealed trait AliveOutputState extends OutputState {
-  val canPush = true
-}
-
-object OutputState{
-
-  case object Dequeueing extends AliveOutputState
-  case class  Streaming(source: Source[DataBuffer], dataQ: DataQueue, post : QueuedItem.PostWrite) extends AliveOutputState
-  case object Suspended extends OutputState {
-    def canPush = true
-  }
-  case object Terminated extends OutputState {
-    def canPush = false
-  }
-
-}
-
-/**
- * This is thrown anytime we hit a state that shouldn't be possible.  If this
- * is ever thrown, there is a bug!
- */
-class InvalidOutputStateException(message: String) extends Exception(message)
-
-
 /** An ADT representing the result of a pushing a message to write
 */
 sealed trait OutputResult
@@ -118,89 +54,77 @@ object OutputResult {
   case class Cancelled(reason: Throwable) extends OutputResult with OutputError
 }
 
-/**
- * The OutputController maintains all state dealing with writing messages in a
- * controller.  It maintains a queue of messages pending write and properly
- * handles writing both regular messages as well as streams.
- */
-trait OutputController[Input, Output] { this : MasterController[Input, Output] =>
-  import OutputState._
+abstract class StaticOutState(val canPush: Boolean) {
+  def disconnecting = !canPush
+}
+object StaticOutState {
+  case object Suspended extends StaticOutState(true)
+  case object Alive extends StaticOutState(true)
+  case object Disconnecting extends StaticOutState(false)
+  case object Terminated extends StaticOutState(false)
+}
 
-  private[controller] var outputState: OutputState = Suspended
-  private var msgq = new MessageQueue[Output](controllerConfig.outputBufferSize)
+trait StaticOutputController[E <: Encoding] extends BaseStaticController[E]{this: ControllerIface[E] =>
 
-  //TODO : This functionality is only used for (and only makes sense) the
-  //ServiceClient.  It probably makes more sense to let ServiceClient do its own
-  //buffering
+
+  private var state: StaticOutState = StaticOutState.Suspended
+  private def disconnecting = state.disconnecting
   private var _writesEnabled = true
+  private var outputBuffer = new MessageQueue[E#Output](controllerConfig.outputBufferSize)
 
-  private var manualDisconnect = false
+  protected def pendingBufferSize = outputBuffer.size
 
-  private[controller] def outputOnConnected() {
+  def writesEnabled = _writesEnabled
 
-    outputState = Dequeueing
+  def pauseWrites() {
+    _writesEnabled = false
+  }
+
+  /**
+   * Resumes writing of messages if currently paused, otherwise has no affect
+   */
+  def resumeWrites() {
     _writesEnabled = true
+    if (!outputBuffer.isEmpty) signalWrite()
+  }
 
-    if (!msgq.isEmpty) {
-      signalWrite()
+  protected def purgePending(reason: Throwable) {
+    while (!outputBuffer.isEmpty) {
+      outputBuffer.dequeue.postWrite(OutputResult.Cancelled(reason))
     }
 
   }
 
-  protected def pendingBufferSize = msgq.size
-
-  private[controller] def outputOnClosed() {
-    val reason = new NotConnectedException("Connection Closed")
-    outputState match {
-      case Streaming(source, dataQ, post) => {
-        source.terminate(reason)
-        post(OutputResult.Failure(reason))
-      }
-      case _ => {}
-    }
-    if (manualDisconnect) {
-      while (!msgq.isEmpty) {
-        msgq.dequeue.postWrite(OutputResult.Cancelled(reason))
-      }
-
-      outputState = Terminated
-    } else {
-      outputState = Suspended
-    }
+  override def connected(endpt: WriteEndpoint) {
+    super.connected(endpt)
+    state = StaticOutState.Alive
+    if (!outputBuffer.isEmpty) signalWrite()
   }
 
-  /**
-   * handle the output side of gracefully disconnecting.  If the output
-   * controller is currently idle, we can immediately switch the state to
-   * Terminated, otherwise, we switch disconnecting to true and wait for pending
-   * messages to either get sent or timeout
-   */
-  private[controller] def outputGracefulDisconnect() {
-    manualDisconnect = true 
-    outputState match {
-      case a : AliveOutputState => {
-        checkOutputGracefulDisconnect()
-      }
-      case other => {
-        val reason = new NotConnectedException("Connection Closed")
-        while (!msgq.isEmpty) {
-          msgq.dequeue.postWrite(OutputResult.Cancelled(reason))
-        }
-        outputState = Terminated
-      }
-    }
+
+  private def onClosed() {
+    if (disconnecting) {
+      val reason = new NotConnectedException("Connection Closed")
+      purgePending(reason)
+    } 
   }
 
-  
-  /**
-   * returns true if the state is switched to Terminated
-   */
-  private def checkOutputGracefulDisconnect(): Boolean = {
-    if (manualDisconnect && msgq.isEmpty && outputState == Dequeueing) {
-      outputState = Terminated
-      checkControllerGracefulDisconnect()
-      true
-    } else false
+  protected def connectionClosed(cause : DisconnectCause) {
+    state = StaticOutState.Terminated
+    onClosed()
+  }
+
+  protected def connectionLost(cause : DisconnectError) {
+    state = StaticOutState.Suspended
+    onClosed()
+  }
+
+  def idleCheck(period: Duration) {
+    val time = System.currentTimeMillis
+    while (!outputBuffer.isEmpty && outputBuffer.head.isTimedOut(time, controllerConfig.sendTimeout)) {
+      val expired = outputBuffer.dequeue
+      expired.postWrite(OutputResult.Cancelled(new RequestTimeoutException))
+    }
   }
 
   private def signalWrite() {
@@ -212,152 +136,51 @@ trait OutputController[Input, Output] { this : MasterController[Input, Output] =
     }
   }
 
-  /** Push a message to be written
-   *
-   * Pushing a message does not necessarily mean it will be written, but rather
-   * that the message is queued to be written.  Messages can be queue
-   * regardless of the state of the underlying connection, even if the
-   * connection is never reconnected.  It is up to the caller to determine
-   * whether a message should be pushed based on connection state.
-   *
-   * @param item the message to push
-   * @param createdMillis the timestamp of when the message was created, defaults to now if not specified
-   * @param postWrite called either when writing has completed or failed
-   *
-   * @return true if the message was successfully enqueued, false if the queue is full
-   */
-  protected def push(item: Output, createdMillis: Long = System.currentTimeMillis)(postWrite: QueuedItem.PostWrite): Boolean = {
+  override def shutdown() {
+    state = StaticOutState.Disconnecting
+    checkShutdown()
+  }
+
+  private def checkShutdown() {
+    if (disconnecting && outputBuffer.isEmpty) {
+      super.shutdown()
+    }
+  }
+
+
+  def canPush = state.canPush && !outputBuffer.isFull
+
+  def push(item: E#Output, createdMillis: Long = System.currentTimeMillis)(postWrite: QueuedItem.PostWrite): Boolean = {
     if (canPush) {
-      if (msgq.size == 0 && outputState == Dequeueing) signalWrite()
-      msgq.enqueue(item, postWrite, createdMillis)
+      if (outputBuffer.isEmpty) signalWrite()
+      outputBuffer.enqueue(item, postWrite, createdMillis)
       true
-    } else {
-      false
-    }
-  }
-
-  protected def canPush = outputState.canPush && !msgq.isFull
-
-  private def drainSource(){ 
-    outputState match {
-      case Streaming(source, dataQ, post) => source.pull {
-        case Success(Some(data)) => {
-          if (dataQ.itemSize == 0) signalWrite()
-          dataQ.enqueue(data)
-          if (! dataQ.isFull) drainSource()
-        }
-        case Success(None) => {
-          //there's actually nothing to do here, switching back to the dequeuing
-          //state is handled in readyForData
-          if (!dataQ.isEmpty) signalWrite()
-        }
-        case Failure(err) => {
-          //todo: where to propagate the error?
-          disconnect()
-        }
-      }
-      case other => throw new InvalidOutputStateException("Attempted to drain source when not in streaming state")
-    }
-  }
-
-
-  /** Purge all pending messages
-   * 
-   * If a message is currently being written, it is not affected
-   */
-  protected def purgePending(reason: Throwable) {
-    while (!msgq.isEmpty) {
-      val q = msgq.dequeue
-      q.postWrite(OutputResult.Cancelled(reason))
-    }
-  }
-
-  protected def writesEnabled: Boolean = _writesEnabled
-
-  /** Purge both pending and outgoing messages */
-  /**
-   * Pauses writing of the next item in the queue.  If there is currently a
-   * message in the process of writing, it will be unaffected.  New messages
-   * can still be pushed to the queue as long as it is not full
-   */
-  protected def pauseWrites() {
-    _writesEnabled = false
-  }
-
-  /**
-   * Resumes writing of messages if currently paused, otherwise has no affect
-   */
-  protected def resumeWrites() {
-    _writesEnabled = true
-    if (!msgq.isEmpty) signalWrite()
+    } else false
   }
 
   def readyForData(buffer: DataOutBuffer) =  {
-    if (checkOutputGracefulDisconnect()) {
+    if (disconnecting && outputBuffer.isEmpty) {
+      checkShutdown()
       MoreDataResult.Complete
-    } else outputState match {
-      case Dequeueing => {
-        while (writesEnabled && outputState == Dequeueing && msgq.size > 0 && ! buffer.isOverflowed) {
-          val next = msgq.dequeue
-          codec.encode(next.item) match {
-            case e: Encoder => {
-              e.encode(buffer)
-              next.postWrite(OutputResult.Success)
-            }
-            case DataStream(source) => {
-              outputState = Streaming(source, new DataQueue(1000), next.postWrite)
-              drainSource()
-            }
-          }
-        }
+    } else {
+      while (writesEnabled && outputBuffer.size > 0 && ! buffer.isOverflowed) {
+        val next = outputBuffer.dequeue
+        codec.encode(next.item, buffer)
+        next.postWrite(OutputResult.Success)
       }
-      case ls @ Streaming(source, dataQ, post) => {
-        while (dataQ.itemSize > 0 && ! buffer.isOverflowed) {
-          dataQ.dequeue.encode(buffer)
-        }
-        if (dataQ.isEmpty) {
-          if (source.isClosed) {
-            //all done
-            post(OutputResult.Success)
-            outputState = Dequeueing
-          } else {
-            //ask for more
-            drainSource()
-          }
-        }
+      if (disconnecting || (writesEnabled && outputBuffer.size > 0)) {
+        //return incomplete only if we overflowed the buffer and have more in
+        //the queue, or if there's nothing left but we're disconnecting to
+        //finish the disconnect process
+        MoreDataResult.Incomplete
+      } else {
+        MoreDataResult.Complete
       }
-      case _ => {}
-    }
-
-    outputState match {
-      case Dequeueing => {
-        if (manualDisconnect || (writesEnabled && msgq.size > 0)) {
-          //return incomplete only if we overflowed the buffer and have more
-          //in the queue, or if disconnecting to finish the disconnect
-          //process
-          MoreDataResult.Incomplete
-        } else {
-          MoreDataResult.Complete
-        }
-      }
-      case Streaming(_, dataQ, _) => {
-        if (manualDisconnect || !dataQ.isEmpty) {
-          MoreDataResult.Incomplete
-        } else {
-          MoreDataResult.Complete
-        }
-      }
-      case _ => MoreDataResult.Complete
     }
   }
         
-  def idleCheck(period: Duration): Unit = {
-    val time = System.currentTimeMillis
-    while (!msgq.isEmpty && msgq.head.isTimedOut(time, controllerConfig.sendTimeout)) {
-      val expired = msgq.dequeue
-      expired.postWrite(OutputResult.Cancelled(new RequestTimeoutException))
-    }
-  }
+      
+
+  
 
 }
-
