+++ colossus/src/main/scala/colossus/core/DataBuffer.scala
@@ -6,18 +6,7 @@ import java.nio.ByteBuffer
 import java.nio.channels.SocketChannel
 
 
-
-
-/**
- * A DataReader is the result of codec's encode operation.  It can either
- * return a DataBuffer, which contains the entire encoded object at once, or it
- * can return a DataStream, which has a Sink for streaming the encoded object.
- */
-sealed trait DataReader
-
-case class DataStream(source: controller.Source[DataBuffer]) extends DataReader
-
-trait Encoder extends DataReader{
+trait Encoder {
   def encode(out: DataOutBuffer)
 
   def bytes: ByteString = {
