+++ colossus/src/main/scala/colossus/protocols/http/HttpResponse.scala
@@ -2,7 +2,6 @@ package colossus
 package protocols.http
 
 import akka.util.{ByteString, ByteStringBuilder}
-import colossus.controller.Source
 import colossus.core.DataBuffer
 import core._
 import service.Callback
@@ -77,32 +76,8 @@ object HttpResponseHead{
 }
   
 
-sealed trait BaseHttpResponse { 
 
-  def head: HttpResponseHead
-
-  /**
-   * Resolves the body into a bytestring.  This operation completes immediately
-   * for static responses.  For streaming responses, this only completes when
-   * all data has been received.  Be aware if the response is an infinite
-   * stream of data (using chunked transfer encoding), this will never
-   * complete.
-   */
-  def resolveBody(): Option[Callback[ByteString]]
-
-  //def withHeader(key: String, value: String): self.type
-
-  def toReader: DataReader
-
-}
-
-//TODO: We need to make some headers like Content-Length, Transfer-Encoding,
-//first-class citizens and separate them from the other headers.  This would
-//prevent things like creating a response with the wrong content length
-
-
-
-case class HttpResponse(head: HttpResponseHead, body: HttpBody) extends BaseHttpResponse with Encoder {
+case class HttpResponse(head: HttpResponseHead, body: HttpBody) extends Encoder {
 
   def encode(buffer: DataOutBuffer) = encode(buffer, HttpHeaders.Empty)
 
@@ -118,15 +93,10 @@ case class HttpResponse(head: HttpResponseHead, body: HttpBody) extends BaseHttp
     body.encode(buffer)
   }
 
-  def resolveBody(): Option[Callback[ByteString]] = if (body.size > 0) {
-    Some(Callback.successful(body.bytes))
-  } else None
 
   def withHeader(key: String, value: String) = copy(head = head.withHeader(key,value))
 
   def code = head.code
-
-  def toReader = this
 
 }
 
@@ -156,53 +126,5 @@ object HttpResponse {
   }
 
 }
-
-
-/**
- * Be aware, at the moment when the response is encoded, there is no processing
- * on the response body, and no headers are added in.  This means if the
- * transfer-encoding is "chunked", the header must already exist and the stream
- * must already be prepending chunk headers to the data.
- */
-case class StreamingHttpResponse(head: HttpResponseHead, body: Option[Source[DataBuffer]]) extends BaseHttpResponse {
-
-  type Encoded = DataReader
-
-  def toReader : DataReader = {
-    val builder = new DynamicOutBuffer(100, false)
-    head.encode(builder)
-    builder write NEWLINE_ARRAY
-
-    val headerBytes = builder.data
-    body.map{stream => 
-      DataStream(new DualSource[DataBuffer](Source.one(headerBytes), stream))
-    }.getOrElse(headerBytes)
-
-  }
-
-  def resolveBody: Option[Callback[ByteString]] = body.map{data => 
-    data.fold(new ByteStringBuilder){(buffer: DataBuffer, builder: ByteStringBuilder) => builder.putBytes(buffer.takeAll)}.map{_.result}
-  }
-
-  def withHeader(key: String, value: String) = copy(head = head.withHeader(key,value))
-
-}
-
-object StreamingHttpResponse {
-
-  def fromStatic(resp: HttpResponse): StreamingHttpResponse = {
-    StreamingHttpResponse(resp.head.withHeader(HttpHeaders.ContentLength, resp.body.size.toString), Some(Source.one(DataBuffer(resp.body.bytes))))
-  }
-
-  def apply[T : HttpBodyEncoder](version : HttpVersion, code : HttpCode, headers : HttpHeaders, data : T) : StreamingHttpResponse = {
-    fromStatic(HttpResponse(
-      HttpResponseHead(version, code, headers), 
-      HttpBody(data)
-    ))
-  }
-
-}
-
-
