+++ colossus/src/main/scala/colossus/protocols/websocket/package.scala
@@ -5,6 +5,7 @@ import core._
 import service._
 import akka.util.{ByteString, ByteStringBuilder}
 import java.util.Random
+import controller.StaticCodec
 
 /**
  * **This package is experimental and subject to breaking changes between
@@ -26,14 +27,16 @@ import java.util.Random
  */
 package object websocket {
 
-  class WebsocketCodec extends Codec[Frame, Frame]{
+  class WebsocketCodec extends StaticCodec[Websocket#ServerEncoding]{
     
     private val random = new Random
     private val parser = FrameParser.frame
 
     def decode(data: DataBuffer) = parser.parse(data)
 
-    def encode(f: Frame) = f.encode(random)
+    def encode(f: Frame, buffer: DataOutBuffer) { buffer write f.encode(random) }
+
+    def endOfStream = None
 
     def reset(){}
   }
